% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finals.R
\name{AmesA}
\alias{AmesA}
\title{Estimate mesophyll surface area to leaf area ratio for each type of mesophyll tissue. To estimate total surface area of mesophyll per leaf area, sum the results of AmesA function for each tissue type.}
\usage{
AmesA(SA, Tt, ASF, V)
}
\arguments{
\item{SA}{number representing the surface area of a mesophyll cell from a specific cell type (eg. spongey mesophyll cell). generated from one of the package's SA functions}

\item{Tt}{number representing the thickness of a mesophyll tissue (eg. thickness of spongey mesophyll). measured}

\item{ASF}{number representing air space fraction for specific mesophyll tissue. estimated by package's ASF function}

\item{V}{number representing volume of specific mesophyll tissue. generated from one of the package's V functions}
}
\value{
number
}
\description{
Estimate mesophyll surface area to leaf area ratio for each type of mesophyll tissue. To estimate total surface area of mesophyll per leaf area, sum the results of AmesA function for each tissue type.
}
\examples{
SA<-685979.3 #spongey mesophyll cell surface area calculated by SAsponge (um^2)
Tt<-126.034 #spongey mesophyll tissue thickness, measured (um)
ASF<-0.3378385 #spongey mesophyll airspace fraction calculated by ASF
V<-53424483 #spongey mesophyll cell volume calculated by Vsponge (um^3)
AmesA(SA,Tt,ASF,V)
}
