% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palisadeH.R
\name{SAh}
\alias{SAh}
\title{Estimate surface area of palisade mesophyll cell in tissue dominated by H-shaped cells}
\usage{
SAh(Dc, Hc, Ds, Hs, j)
}
\arguments{
\item{Dc}{numerical value representing the diameter (or mean) of the central cylinder of the H-shaped palisade mesophyll cell}

\item{Hc}{numerical value representing the height (or mean) of the central cylinder of the H-shaped palisade mesophyll cell}

\item{Ds}{numerical value representing the diameter (or mean) of one side cylinder of the H-shaped palisade mesophyll cell}

\item{Hs}{numerical value representing the height (or mean) of one side cylinder of the H-shaped palisade mesophyll cell}

\item{j}{numerical value representing the total number of arms (aka side cylinders) in the palisade mesophyll H cell}
}
\value{
number
}
\description{
Estimate surface area of palisade mesophyll cell in tissue dominated by H-shaped cells
}
\examples{
Ds<-mean(18.04, 18.67, 17.19)
Hs<-mean(34.65, 37.48, 29.90)
Dc<-mean(13.55, 17.91, 21.12)
Hc<-mean(38.71, 49.27, 41.85)
j<-2
SAh(Dc,Hc,Ds,Hs, j)
}
