% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palisadeI.R
\name{SAi}
\alias{SAi}
\title{Estimate surface area of palisade mesophyll cell in tissue dominated by I-shaped cells}
\usage{
SAi(Di, Hi)
}
\arguments{
\item{Di}{numerical value representing the diameter of the I-shaped (aka cylindrical) palisade mesophyll cell (or average thereof)}

\item{Hi}{numerical value representing height (or average)of the I-shaped palisade mesophyll cell. should be measured from same cell(s) as Di}
}
\value{
a number
}
\description{
Estimate surface area of palisade mesophyll cell in tissue dominated by I-shaped cells
}
\examples{
Di<-mean(18.04, 18.67, 17.19)
Hi<-mean(34.65, 37.48, 29.90)
SAi(Di, Hi)

}
