% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spongey.R
\name{SAsponge}
\alias{SAsponge}
\title{Estimate surface area of spongey mesophyll cell}
\usage{
SAsponge(Ps)
}
\arguments{
\item{Ps}{Perimeter of spongey mesophyll cell (numerical) or average perimeter of cells}
}
\value{
A number
}
\description{
Estimate surface area of spongey mesophyll cell
}
\examples{
Ps<-mean(153.533, 184.481, 206.374)
SAsponge(Ps)

}
