% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palisadeI.R
\name{Vi}
\alias{Vi}
\title{Estimate volume of palisade mesophyll cell in tissue dominated by I-shaped cells}
\usage{
Vi(Di, Hi)
}
\arguments{
\item{Di}{numerical value representing diameter (or average) of the I-shaped palisade mesophyll cell}

\item{Hi}{numerical value representing height (or average) of the I-shaped palisade mesophyll cell. should be measured from same cell(s) as Di}
}
\value{
number
}
\description{
Estimate volume of palisade mesophyll cell in tissue dominated by I-shaped cells
}
\examples{
Di<-mean(18.04, 18.67, 17.19)
Hi<-mean(34.65, 37.48, 29.90)
Vi(Di, Hi)
}
