% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spongey.R
\name{Vsponge}
\alias{Vsponge}
\title{Estimate volume of spongey mesophyll cell}
\usage{
Vsponge(Ps)
}
\arguments{
\item{Ps}{Perimeter of spongey mesophyll cell (numerical) or average perimeter of cells}
}
\value{
A number
}
\description{
Estimate volume of spongey mesophyll cell
}
\examples{
Ps<-mean(153.533, 184.481, 206.374)
Vsponge(Ps)
}
